--- src/main/java/apollosupport/federation/FieldRequest.java
@@ -1,5 +1,0 @@
-package apollosupport.federation;
-
-public class FieldRequest {
-
-}
