+++ src/main/java/com/intuit/graphql/orchestrator/apollofederation/DefaultDataFetcher.java
@@ -1,4 +1,4 @@
-package apollosupport.federation;
+package com.intuit.graphql.orchestrator.apollofederation;
 
 import graphql.GraphQLContext;
 import graphql.schema.DataFetcher;
@@ -6,16 +6,10 @@ import graphql.schema.DataFetchingEnvironment;
 import lombok.Data;
 
 /**
- * This {@link DataFetcher} is used for sending entity data request to a subgraph using
- * <pre>
- * {@code
- * query {
- *    _entities(...) {...}}
- * }
- * </pre>
+ * This {@link DataFetcher} is used to fetch fields for an Apollo Federation compliant sub-graph
  */
 @Data
-public class EntityDataFetcher implements DataFetcher<Object> {
+public class DefaultDataFetcher implements DataFetcher<Object> {
 
   private SubGraphContext subGraphContext;
