+++ src/main/java/com/intuit/graphql/orchestrator/apollofederation/FieldRequest.java
@@ -1,0 +1,5 @@
+package com.intuit.graphql.orchestrator.apollofederation;
+
+public class FieldRequest {
+
+}
